Fix typos in GitSCMFileSystemTest naming
The given_filesystem_when_askingChangesSinceNewRevision_then_changesArePopulatedButEmpty
test had a confusing mix of xxx261 and xxx260 variables that were
referring to tags which sometimes did not match the name of the variable.

Makes the names consistent rather than risk confusing the reader of the
test.

Other tests referred to a GitSCMFileSystem instance as "instance" when
it was then used as a file system reference to a specific commit.
Changed the name to be clearer where it was referring.

Preparing to remove the requirement that tests require all tags and
all history in the working directory.  Optimizations in git plugin
3.4.0 remove that assumption (and its resulting waste of time and
disc space).